refactor(repl): move add_as_custom_prim to HelperCommandSet
Puts the command in the module matching its Helper category and drops the
now-unused imports from cmd_fabric_gen.
